Decomponentify
