+++ src/lib/components/Modal.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
-	export let dialog: HTMLDialogElement | null = null;
+	export let dialog: HTMLDialogElement;
 </script>
 
-<dialog bind:this={dialog} on:close>
+<dialog bind:this={dialog}>
 	<slot />
 </dialog>
